5758 - Add NationalDataPointValidations.hasError
