+++ src/meta/assessment/validation/nationalDataPointValidations/index.ts
@@ -1,5 +1,7 @@
 import { calculateSummary } from 'meta/assessment/validation/nationalDataPointValidations/calculateSummary'
+import { hasError } from 'meta/assessment/validation/nationalDataPointValidations/hasError'
 
 export const NationalDataPointValidations = {
   calculateSummary,
+  hasError,
 }
